# docs/coze-plugin-openapi-v2.yaml
openapi: "3.0.0"
info:
  title: 数学教学智能体
  description: 数学作业批改、智能出题、学习诊断、OCR识别
  version: "1.0.0"
servers:
  - url: https://5ee709a4e70d57.lhr.life
paths:
  /plugin/ocr:
    post:
      summary: OCR识别
      description: 从图片中提取文字
      operationId: ocrRecognize
      parameters: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - image_url
              properties:
                image_url:
                  type: string
                  description: 图片URL
      responses:
        "200":
          description: 识别结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  text:
                    type: string

  /plugin/grade:
    post:
      summary: 批改作业
      description: AI自动批改数学作业，返回评分和批改详情
      operationId: gradeHomework
      parameters: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - student_name
                - questions_and_answers
              properties:
                student_name:
                  type: string
                  description: 学生姓名
                school_level:
                  type: string
                  description: 学段（小学/初中/高中）
                  default: 初中
                questions_and_answers:
                  type: string
                  description: 作业题目和答案
      responses:
        "200":
          description: 批改结果

  /plugin/generate-exam:
    post:
      summary: 智能出题
      description: 根据知识点和难度生成试卷
      operationId: generateExam
      parameters: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - school_level
              properties:
                school_level:
                  type: string
                  description: 学段（小学/初中/高中）
                knowledge_points:
                  type: string
                  description: 薄弱知识点（逗号分隔）
                difficulty:
                  type: integer
                  description: 难度（1-5）
                  default: 3
                question_count:
                  type: integer
                  description: 题目数量
                  default: 10
      responses:
        "200":
          description: 生成的试卷

  /plugin/diagnose:
    post:
      summary: 学习诊断
      description: 根据学生表现生成诊断报告
      operationId: diagnoseStudent
      parameters: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - student_name
                - performance_data
              properties:
                student_name:
                  type: string
                  description: 学生姓名
                school_level:
                  type: string
                  description: 学段
                  default: 初中
                performance_data:
                  type: string
                  description: 近期考试或作业表现数据
      responses:
        "200":
          description: 诊断报告

  /plugin/learning-plan:
    post:
      summary: 学习计划
      description: 根据薄弱知识点生成学习计划
      operationId: learningPlan
      parameters: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - student_name
                - weak_points
              properties:
                student_name:
                  type: string
                  description: 学生姓名
                school_level:
                  type: string
                  description: 学段
                  default: 初中
                weak_points:
                  type: string
                  description: 薄弱知识点（逗号分隔）
      responses:
        "200":
          description: 学习计划

  /plugin/knowledge-point:
    get:
      summary: 知识点查询
      description: 查询标准知识点信息
      operationId: queryKnowledgePoint
      parameters:
        - name: text
          in: query
          required: true
          schema:
            type: string
          description: 知识关键词
        - name: level
          in: query
          schema:
            type: string
            default: 初中
          description: 学段
      responses:
        "200":
          description: 知识点信息

  /plugin/health:
    get:
      summary: 健康检查
      description: 检查服务状态
      operationId: checkHealth
      responses:
        "200":
          description: 服务状态